Don't scroll main view to top after adding or deleting a transaction
This was implemented assuming an added transaction would likely appear
at the top but just keeping the scroll position is probably the better
option. There are also cases where the user checks if a certain
has been already added and if not, adds it, which then appears right at
the original scroll position.

fixes #205