Trim new participant names (fixes issue that created duplicates)
It's probably generally better to trim the names.

An issue that occurred because of a trailing space was that creating
a direct transaction lead to duplicate participants. It happened because
the names were set as child of each <option> element which effectively
trims it:

<select>
  <option>Participant Name </option>
</select>
select.value => "Participant Name"

This is fixed as well by using the `value` attribute.